refactor: exports from service/util
